Refactor: move RelaxedPathFilter into separate file
